added regression test for queue copy
