+++ raiden/network/protocol.py
@@ -9,7 +9,6 @@ from itertools import repeat
 
 import cachetools
 import gevent
-from gevent.queue import Queue
 from gevent.event import (
     _AbstractLinkable,
     AsyncResult,
@@ -34,6 +33,7 @@ from raiden.settings import (
 )
 from raiden.messages import decode, Ack, Ping, SignedMessage
 from raiden.utils import isaddress, sha3, pex
+from raiden.utils.notifying_queue import NotifyingQueue
 
 log = slogging.get_logger(__name__)  # pylint: disable=invalid-name
 
@@ -360,39 +360,6 @@ def healthcheck(
                 receiver_address,
                 NODE_NETWORK_REACHABLE,
             )
-
-
-class NotifyingQueue(Event):
-    def __init__(self):
-        super(NotifyingQueue, self).__init__()
-        self._queue = Queue()
-
-    def put(self, item):
-        """ Add new item to the queue. """
-        self._queue.put(item)
-        self.set()
-
-    def get(self, block=True, timeout=None):
-        """ Removes and returns an item from the queue. """
-        value = self._queue.get(block, timeout)
-        if self._queue.empty():
-            self.clear()
-        return value
-
-    def peek(self, block=True, timeout=None):
-        return self._queue.peek(block, timeout)
-
-    def __len__(self):
-        return len(self._queue)
-
-    def copy(self):
-        """ Copies the current queue items. """
-        copy = self._queue.copy()
-
-        result = list()
-        while not copy.empty():
-            result.append(copy.get_nowait())
-        return result
 
 
 class RaidenProtocol(object):
